requirements.txt
streamlit==1.29.0
google-generativeai>=0.3.2
SpeechRecognition==3.10.0
gTTS==2.4.0
googletrans==4.0.0-rc1
Pillow>=10.0.1
python-dotenv>=1.0.0